Refactor Solana components for improved organization and maintainability
- Updated import paths for ClusterProvider and WalletButton to consolidate Solana-related components under a unified directory structure.
- Removed obsolete Solana provider and cluster data access components to streamline the codebase.
- Enhanced the layout by integrating user wallet components and cluster management into a more cohesive structure, improving overall code organization.
