+++ src/components/solana/user-wallet/wallet-connect.tsx
@@ -12,9 +12,9 @@ import {
 } from '@solana/wallet-adapter-react'
 import { WalletModalProvider } from '@solana/wallet-adapter-react-ui'
 import { ReactNode, useCallback, useEffect, useMemo } from 'react'
-import { useCluster } from './cluster/cluster-data-access'
+import { useCluster } from '../../../solana/clusters/cluster-data-access'
 import toast from 'react-hot-toast'
-import { validateRpcConnection } from '@/utils/solana/validateRpcConnection'
+import { validateRpcConnection } from '@/solana/utils/validateRpcConnection'
 import { clusterApiUrl } from '@solana/web3.js'
 
 require('@solana/wallet-adapter-react-ui/styles.css')
